Change the integration interval start point
This changes the integration interval to start at the start point and
spreads the interval out in logaritmic space in a way in which the
distance to the firt point can be directly specified.  Currently that
is hard-coded.  Probably should change the meaning of the parameters
slighty to make this behave in a more useful way. (i.e. make
the integration start at zero, since I don't think anything else is
really meaningful for the ZND, and make x_start the location of the
first point.)