Don't renorm abundances in jacobian evaluation
The jacobian evaluation was silently renorming the integrated variables
(the abundances) which could lead to some rather odd integration behavior.
The reason this was being done was to protect from passing abundances that
don't sum to 1 to the mesa routines.  Now I just use a local copy of the
abundances that is renormed to pass to the mesa routines without messing with
the integration variables passed in.
Note that the sum to 1 can get bad near the singularity in the ZND equations.
There is still an outstanding problem with that due to how the singularity
is detected.  That issue has not yet been fixed, and is currently preventing
the finding of the eigenvalue solution for CO detonations above a density
of about 8e6 g/cc.